Count in 32 bits so the counters offload on GFX9
The GPU offload has no 64-bit atomic on GFX9 - global_atomic_add_x2 is
refused there - and a percpu counter it cannot add to atomically it cannot
offload at all: 128 lanes would read, add and write it back and keep one.
So the verifier turns the program away with a 64-bit lb_stats.

Narrow the two fields to 32 bits and widen on the host, which already sums
the per-instance values into a uint64 and now just casts on the way in.
Nothing above manager.go sees a difference.

At 45 Mpps a packet count wraps in about 95 seconds.  The read path sums
whatever is there when asked, so a reader that looks less often than that
will see a total go backwards; that is the cost of this, and it is noted
next to the struct.

+++ internal/lb/manager.go
@@ -356,8 +356,8 @@ func (m *Manager) GetStats(vip model.VIP) (*model.StatsEntry, error) {
 
 	var packets, bytes uint64
 	for i := range perCPU {
-		packets += perCPU[i].V1
-		bytes += perCPU[i].V2
+		packets += uint64(perCPU[i].V1)
+		bytes += uint64(perCPU[i].V2)
 	}
 	return &model.StatsEntry{
 		Packets: packets,
@@ -476,8 +476,8 @@ func (m *Manager) GetGlobalStats() map[string]model.StatsEntry {
 		// there to read.
 		var e model.StatsEntry
 		for i := range perCPU {
-			e.Packets += perCPU[i].V1
-			e.Bytes += perCPU[i].V2
+			e.Packets += uint64(perCPU[i].V1)
+			e.Bytes += uint64(perCPU[i].V2)
 		}
 		result[name] = e
 	}
